찐찐막 최종데모

+++ app/src/main/java/com/jubging/jubging/ui/jubging/BluetoothActivity.kt
@@ -123,7 +123,7 @@ class BluetoothActivity: AppCompatActivity(),PictureView {
                 this.intent.getIntExtra("walk",0).let { intent.putExtra("walk", it)}
                 this.intent.getStringExtra("time").let { intent.putExtra("time", it)}
                 this.intent.getFloatExtra("distance",0.0f).let { intent.putExtra("distance", it)}
-                this.intent.getFloatExtra("kcal",0.0f).let { intent.putExtra("kcal", it)}
+                this.intent.getIntExtra("kcal",0).let { intent.putExtra("kcal", it)}
                 this.intent.getBooleanExtra("tongs_return",true).let { intent.putExtra("tongs_return", it)}
                 this.intent.getIntExtra("jubjubi_id",0).let { intent.putExtra("jubjubi_id", it)}
                 this.intent.getIntExtra("tongs_id",0).let { intent.putExtra("tongs_id", it)}
