+++ app/src/main/java/com/jubging/jubging/ui/jubging/ShareActivity.kt
@@ -58,7 +58,7 @@ class ShareActivity: BaseActivity<ActivityShareBinding>(ActivityShareBinding::in
         this.intent.getStringExtra("WEIGHT")?.let { Log.d("SHARE", it) }
         binding.shareWeightDataTv.text = this.intent.getStringExtra("WEIGHT")
         binding.shareDistanceDataTv.text = this.intent.getFloatExtra("distance",0.0f).toString()
-        binding.shareKcalDataTv.text = this.intent.getFloatExtra("kcal",0.0f).toString()
+        binding.shareKcalDataTv.text = this.intent.getIntExtra("kcal",0).toString()
         binding.shareTimeDataTv.text = this.intent.getStringExtra("time").toString()
         binding.shareWalkDataTv.text = this.intent.getIntExtra("walk",0).toString()
         //다운로드 버튼 클릭시
